# static/content/openapi.yaml
openapi: 3.1.0
info:
  title: Chronoverse HTTP API
  version: 1.0.0
  description: Public session-authenticated API exposed by the Chronoverse HTTP gateway.
servers:
  - url: http://localhost:8080
    description: Development
  - url: /api
    description: Production Nginx proxy
...
paths:
...
  /workflows:
    get:
      operationId: listWorkflows
      summary: List workflows
      tags: [Workflows]
      security: [{cookieSession: []}]
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {name: query, in: query, schema: {type: string}, description: Workflow name search}
        - {name: kind, in: query, schema: {type: string, enum: [HEARTBEAT, CONTAINER]}}
        - {name: build_status, in: query, schema: {$ref: '#/components/schemas/BuildStatus'}}
        - {name: terminated, in: query, schema: {type: boolean, default: false}}
        - {name: interval_min, in: query, description: Minimum interval in minutes; zero is treated as unset, schema: {type: integer, minimum: 0, maximum: 2147483647}}
        - {name: interval_max, in: query, description: Maximum interval in minutes; zero is treated as unset, schema: {type: integer, minimum: 0, maximum: 2147483647}}
      responses:
        '200':
          description: Paginated workflow list
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflows: {type: array, items: {$ref: '#/components/schemas/WorkflowListItem'}}
                  cursor: {type: string}
        '401': {$ref: '#/components/responses/Unauthorized'}
    post:
      operationId: createWorkflow
      summary: Create a workflow
      tags: [Workflows]
      security: [{cookieSession: [], csrfToken: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WorkflowInput'}
      responses:
        '201':
          description: Workflow created
          content:
            application/json:
              schema:
                {$ref: '#/components/schemas/IdentifierResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
  /workflows/{workflow_id}:
    parameters: [{$ref: '#/components/parameters/WorkflowId'}]
    get:
      operationId: getWorkflow
      summary: Get a workflow
      tags: [Workflows]
      security: [{cookieSession: []}]
      responses:
        '200':
          description: Workflow details
          content: {application/json: {schema: {$ref: '#/components/schemas/WorkflowDetail'}}}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      operationId: updateWorkflow
      summary: Update a workflow
      tags: [Workflows]
      security: [{cookieSession: [], csrfToken: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/WorkflowUpdateInput'}}}
      responses:
        '204': {description: Workflow updated}
        '409': {$ref: '#/components/responses/Conflict'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      operationId: terminateWorkflow
      summary: Terminate a workflow
      tags: [Workflows]
      security: [{cookieSession: [], csrfToken: []}]
      responses:
        '204': {description: Workflow terminated}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
    delete:
      operationId: deleteWorkflow
      summary: Delete a workflow
      tags: [Workflows]
      security: [{cookieSession: [], csrfToken: []}]
      responses:
        '204': {description: Terminated workflow deleted}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: session
      description: Encrypted session cookie issued after registration or login.
    csrfToken:
      type: apiKey
      in: cookie
      name: csrf
      description: CSRF cookie issued with the session and required for state-changing requests.
  parameters:
    WorkflowId: {name: workflow_id, in: path, required: true, schema: {type: string, format: uuid}}
...
    Cursor: {name: cursor, in: query, required: false, schema: {type: string}, description: Opaque pagination cursor}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Valid UTF-8 containing 1-255 bytes after trimming only outer ASCII spaces (U+0020), with no Unicode control characters. Reuse a key only for retries of the exact same logical request.
      schema:
        type: string
        minLength: 1
        x-maxBytes: 255
  responses:
    BadRequest: {description: Invalid input or missing required idempotency key}
    Conflict: {description: Idempotency key was already used with a different logical request}
    Unauthorized: {description: Missing or invalid session}
    NotFound: {description: Resource not found}
    PreconditionFailed: {description: Current workflow or job state does not allow the operation}
...
  schemas:
...
    WorkflowInput:
      type: object
      required: [name, payload, kind, interval, max_consecutive_job_failures_allowed]
      properties:
        name: {type: string}
        payload: {type: string, description: Workflow-kind-specific JSON encoded as a string}
        kind: {type: string, enum: [HEARTBEAT, CONTAINER]}
        interval: {type: integer, minimum: 1, maximum: 10080, description: Schedule interval in minutes}
        max_consecutive_job_failures_allowed: {type: integer, minimum: 3, maximum: 100}
        log_retention: {type: boolean, description: Supported for CONTAINER workflows; forced off for HEARTBEAT}
    WorkflowUpdateInput:
      type: object
      required: [name, payload, interval, max_consecutive_job_failures_allowed]
      properties:
        name: {type: string}
        payload: {type: string, description: Workflow-kind-specific JSON encoded as a string}
        interval: {type: integer, minimum: 1, maximum: 10080, description: Schedule interval in minutes}
        max_consecutive_job_failures_allowed: {type: integer, minimum: 3, maximum: 100}
    BuildStatus: {type: string, enum: [QUEUED, STARTED, COMPLETED, FAILED, CANCELED]}
...
    IdentifierResponse:
      type: object
      required: [id]
      properties:
        id: {type: string, format: uuid}
    WorkflowListItem:
      type: object
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        payload: {type: string, description: Workflow-kind-specific JSON encoded as a string}
        kind: {type: string, enum: [HEARTBEAT, CONTAINER]}
        build_status: {$ref: '#/components/schemas/BuildStatus'}
        interval: {type: integer, description: Schedule interval in minutes}
        consecutive_job_failures_count: {type: integer, minimum: 0}
        max_consecutive_job_failures_allowed: {type: integer, minimum: 3, maximum: 100}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        terminated_at: {type: string, format: date-time}
        log_retention: {type: boolean}
    WorkflowDetail:
      allOf:
        - {$ref: '#/components/schemas/WorkflowListItem'}
        - type: object
          properties:
            generation: {type: integer, minimum: 0}
...